+++ packages/ai/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
-- Fixed Xiaomi MiMo Anthropic-compat endpoints (`*.xiaomimimo.com/anthropic`, including every Token Plan region and `api.xiaomimimo.com`) losing prior-turn reasoning on continuation requests. `convertAnthropicMessages` treated all unknown endpoints as signing-capable and demoted MiMo's unsigned `thinking` blocks to `type: "text"`, which destabilized tool-call argument serialization on the next turn — the upstream symptom behind the `args?.ops?.map is not a function` crash reported against the `todo` tool ([#2005](https://github.com/can1357/oh-my-pi/issues/2005)). `isNonSigningAnthropicEndpoint` now recognizes the Xiaomi family (provider id `xiaomi` / `xiaomi-token-plan-*` and the `xiaomimimo.com` host suffix) so unsigned thinking blocks replay as `type: "thinking"`.
+- Fixed Anthropic-compatible reasoning endpoints losing prior-turn reasoning on continuation requests when they emit unsigned `thinking` blocks. `convertAnthropicMessages` treated unknown endpoints as signature-enforcing and demoted unsigned reasoning to `type: "text"`, which destabilized tool-call argument serialization on the next turn — the upstream symptom behind the `args?.ops?.map is not a function` crash reported against the `todo` tool. Official `api.anthropic.com` keeps the conservative text fallback; non-official `anthropic-messages` reasoning models now replay unsigned reasoning as native `type: "thinking"` ([#2005](https://github.com/can1357/oh-my-pi/issues/2005)).
 
 ## [15.9.67] - 2026-06-06
